完成除WEB和上传外的所有功能开发和测试；增加CLI启动项目；增加弹幕、SC、舰队、礼物录制和对应的重连机制以及配置；对增加的API进行缓存建档；增加对弹幕录制机制进行下载逻辑处理的代码；

+++ DDTV_Core/SystemAssembly/BilibiliModule/API/LiveChatScript/EntryEffectEventArgs.cs
@@ -16,7 +16,7 @@ namespace DDTV_Core.SystemAssembly.BilibiliModule.API.LiveChatScript
         /// <summary>
         /// 舰长uid
         /// </summary>
-        public int uid { set; get; }
+        public long uid { set; get; }
         /// <summary>
         /// 欢迎信息
         /// </summary>
@@ -28,7 +28,7 @@ namespace DDTV_Core.SystemAssembly.BilibiliModule.API.LiveChatScript
 
         internal EntryEffectEventArgs(JObject obj) : base(obj)
         {
-            uid = (int)obj["data"]["uid"];
+            uid = (long)obj["data"]["uid"];
             face = (string)obj["data"]["face"];
             copy_writing = (string)obj["data"]["copy_writing"];
             copy_writing_v2 = (string)obj["data"]["copy_writing_v2"];
